// client/src/features/auth/authSchema.ts
import { z } from "zod";

export const registerSchema = z.object({
    name: z.string().trim().min(1, "Full name is required"),
    email: z.string().trim().min(1, "Email is required").email("Enter a valid email address"),
    password: z.string().min(1, "Password is required").min(8, "Password must be at least 8 characters")

});


export const loginSchema = z.object({
    email: z.string().trim().min(1, "Email is required").email("Please enter a valid email"),
    password: z.string().min(1, "Password is required"),
})

export type RegisterFormData = z.infer<typeof registerSchema>
export type LoginFormData = z.infer<typeof loginSchema>;